refactor issue metadata parser helpers
